Hashed curves | new delta

+++ lib/Helper_Functions.h
@@ -7,4 +7,4 @@ int Read_curve_files(vector<vector<double*>>*, vector<vector<double*>>*, char*, 
 double point_dist(double*, double*, int);
 int dist(vector<int>*, vector<int>*, int, int=1);
 void brute_force(vector<vector<int>>*, vector<vector<int>>*, vector<int>*, vector<double>*);
-int* arg_min(double**, vector<int>*, int, int);
+double* arg_min(double**, vector<int>*, double, int);
