+++ lib/Traversals.h
@@ -6,4 +6,4 @@ void Traversal(double ***, vector<double*>*, vector<double*>*);
 void DTW(double ***, vector<double*>*, vector<double*>*);
 double min(double, double, double);
 void shift_grid(vector<int>*, int, int);
-void hash_curve(vector<int*>*, vector<double*>*, vector<int>*, int, int);
+void hash_curve(vector<double*>*, vector<double*>*, vector<int>*, double, int);
